Upgrade next and add cursor fetching for thoughts
The Notion API by default only returns 100 children for a block, and because I ahve over 100 thoughts (crazy!), new ones weren't getting picked up. This commit refactors the code to continue sending requests until all thoughts are fetched. This may want to be rethought (heh) if the number of thoughts gets incredibly large, but I think for now this is a fine approach and saves me from having to do any pagination / virtualized list stuff on the client.